requirements.txt
mcp
mcp[cli]
playwright
requests
python-dotenv
Pillow